#. 몬스터 죽고, 실제 재화가 반영되는거 까지 완료
- 재화 매니저 개선
- 보상 매니저 제작
- 유닛 죽었을떄 보상 처리

+++ Assets/02. Scripte/DataLibrary/Main/DataTable/Data_Monster.cs
@@ -58,12 +58,12 @@ namespace IdleGame.Data
         /// <summary>
         /// [데이터] 보상 경험치입니다.
         /// </summary>
-        public float experience_reward = -1;
+        public long experience_reward = -1;
 
         /// <summary>
-        /// [데이터] 보상 경험치입니다.
+        /// [데이터] 보상 재화입니다..
         /// </summary>
-        public float gold_reward = -1;
+        public long gold_reward = -1;
 
         /// <summary>
         /// [데이터] 동료 보상유무를 나타냅니다.
